Explicitly set the number of replicas
If the replicas field is not explicitly set it
defaults to 1.

Although this doesn't seem to be a problem some
Gitops tools (e.g. flux) use 'kubectl apply -f'
under the hood and will not be able to reconcile
the state if someone manually scales down the
deployment and replicas are not explicitly set