fix: allow forward slashes and special characters in context name
Having special characters in the context name broke the navigation
because the special characters where not escaped and broke the router
assumptions both in the frontend and in the backend.

With this fix we escape the special characters in the context name both
in the frontend and when calling the API server in the backend.

Fixes #647

chore: add license

chore: JS linting