chore(ci): fix CI definition for >1 workers
Our drone configuration stopped working when we added some more workers. We were abusing the pipelines to group steps together, but Drone runs each pipeline in a different worker without shared volumes, etc. So we can't use the image build in one worker in the other.

There does not seem to be an easy way of choosing to run the pipelines in the same worker, so we need to refactor all the pipelines.